Squashed commit of the following:

commit 94cf98756403ef8cda5437aa877b92e21dbcbc43
Author: Dominik Tomasiewicz <dominik.tomasiewicz@relativity.com>
Date:   Fri Nov 5 10:11:32 2021 +0000

    Merge pull request #1175 in DTX/import-api-rdc from pipeline-test-REL-604065-develop-update-tests to develop

+++ Source/Relativity.DataExchange.NUnit.Integration/Service/KeplerCodeManagerTests.cs
@@ -24,6 +24,7 @@ namespace Relativity.DataExchange.NUnit.Integration.Service
 	[TestFixture(true)]
 	[TestFixture(false)]
 	[Feature.DataTransfer.ImportApi]
+	[IgnoreIfVersionLowerThan(RelativityVersion.Indigo)] // IChoiceManager does not exist in old versions
 	public class KeplerCodeManagerTests : KeplerServiceTestBase
 	{
 		private const string EmptySingleChoiceFieldName = "EmptySingleChoiceField";
@@ -36,6 +37,7 @@ namespace Relativity.DataExchange.NUnit.Integration.Service
 		private int emptySingleChoiceFieldId;
 		private int populatedSingleChoiceFieldId;
 		private int parentId;
+		private bool testsSkipped;
 
 		private kCura.EDDS.WebAPI.DocumentManagerBase.Field emptySingleChoiceFieldCodeType;
 		private kCura.EDDS.WebAPI.DocumentManagerBase.Field populatedSingleChoiceFieldCodeType;
@@ -88,33 +90,42 @@ namespace Relativity.DataExchange.NUnit.Integration.Service
 		[OneTimeSetUp]
 		public async Task OneTimeSetUpAsync()
 		{
-			this.parentId = await WorkspaceHelper.ReadRootArtifactIdAsync(this.TestParameters, this.TestParameters.WorkspaceId).ConfigureAwait(false);
-
-			this.emptySingleChoiceFieldId = await FieldHelper.CreateSingleChoiceFieldAsync(
-				                                    this.TestParameters,
-				                                    EmptySingleChoiceFieldName,
-				                                    (int)ArtifactType.Document,
-				                                    isOpenToAssociations: false).ConfigureAwait(false);
-
-			this.populatedSingleChoiceFieldId = await FieldHelper.CreateSingleChoiceFieldAsync(
-				                              this.TestParameters,
-				                              PopulatedSingleChoiceFieldName,
-				                              (int)ArtifactType.Document,
-				                              isOpenToAssociations: false).ConfigureAwait(false);
-
-			var choicesValueSource = ChoicesValueSource.CreateSingleChoiceSource(ChoicesCount);
-			await ChoiceHelper.ImportValuesIntoChoiceAsync(
+			this.testsSkipped = RelativityVersionChecker.VersionIsLowerThan(
 				this.TestParameters,
-				this.populatedSingleChoiceFieldId,
-				choicesValueSource).ConfigureAwait(false);
-			this.uniqueValues = choicesValueSource.UniqueValues;
+				RelativityVersion.Indigo);
+			if (!this.testsSkipped)
+			{
+				this.parentId = await WorkspaceHelper.ReadRootArtifactIdAsync(this.TestParameters, this.TestParameters.WorkspaceId).ConfigureAwait(false);
+
+				this.emptySingleChoiceFieldId = await FieldHelper.CreateSingleChoiceFieldAsync(
+					                                this.TestParameters,
+					                                EmptySingleChoiceFieldName,
+					                                (int)ArtifactType.Document,
+					                                isOpenToAssociations: false).ConfigureAwait(false);
+
+				this.populatedSingleChoiceFieldId = await FieldHelper.CreateSingleChoiceFieldAsync(
+					                                    this.TestParameters,
+					                                    PopulatedSingleChoiceFieldName,
+					                                    (int)ArtifactType.Document,
+					                                    isOpenToAssociations: false).ConfigureAwait(false);
+
+				var choicesValueSource = ChoicesValueSource.CreateSingleChoiceSource(ChoicesCount);
+				await ChoiceHelper.ImportValuesIntoChoiceAsync(
+					this.TestParameters,
+					this.populatedSingleChoiceFieldId,
+					choicesValueSource).ConfigureAwait(false);
+				this.uniqueValues = choicesValueSource.UniqueValues;
+			}
 		}
 
 		[OneTimeTearDown]
 		public async Task OneTimeTearDownAsync()
 		{
-			await FieldHelper.DeleteFieldAsync(this.TestParameters, this.populatedSingleChoiceFieldId).ConfigureAwait(false);
-			await FieldHelper.DeleteFieldAsync(this.TestParameters, this.emptySingleChoiceFieldId).ConfigureAwait(false);
+			if (!this.testsSkipped)
+			{
+				await FieldHelper.DeleteFieldAsync(this.TestParameters, this.populatedSingleChoiceFieldId).ConfigureAwait(false);
+				await FieldHelper.DeleteFieldAsync(this.TestParameters, this.emptySingleChoiceFieldId).ConfigureAwait(false);
+			}
 		}
 
 		[IdentifiedTest("6a239075-617b-4c76-97e9-fb28f6effeff")]
